dxva: add support for new dxva2 and d3d11 hwaccel APIs
This also adds support to avconv (which is trivial due to the new
hwaccel API being generic enough). For now, this keeps avconv_dxva2.c as
"dxva2-old", although it doesn't work as avconv.c can't handle multiple
hwaccels with the same pixfmt. It will be removed in a later commit.

The new decoder setup code in dxva2.c is significantly based on work by
Steve Lhomme <robux4@gmail.com>, but with heavy changes/rewrites.